Update Configuration defaults in standalone mode
If you provide kubelet with a KubeletConfiguration file, it will adopt
sensible defaults for normal mode. However these defaults (webhook auth)
do not make sense in standalone mode, and will result in Kubelet
crashing.

We now explictly disable Webhook Authorization and Authentication in the
default configuration if the Bookstrap Provider does not provide an API.